Mockup of capacative zones for pan and twist degrees of freedom
This is just a mockup for comment, not a working board. Enclosed copper
areas should be filled and connected to appropriate nets.

As proposed https://twitter.com/databasescaling/status/1388295587062243331

The boards in the stack would all be the same but with different parts
wired up.

A stack of two boards could provide the two pan degrees of freedom in X and
Y. A stack of three boards could provide the twist degree of freedom as
well.

The middle board would use the radial pad on the top layer and the
rectangular pads on the bottom layer. The top board would use the radial
pads on the bottom layer to complete the twist sensor. The bottom board
would use the rectangular pads on the top layer to complete the X and Y
"pan" sensors.

The boards would be mounted concentrically in the enclosure but bonded to
different parts. The enclosure would have to ensure that they remained a
small and well specified distance apart.

Dimensions are not properly specified yet but hopefully there will be
sufficient space on the top and bottom boards for the supporting circuitry.


Thoughts?


Signed-off-by: Andy Bennett <andyjpb@ashurst.eu.org>